CRUD Order and Order detail

+++ src/main/java/com/locboy/locboybackend/dtos/OrderDetailDTO.java
@@ -28,7 +28,7 @@ public class OrderDetailDTO {
 
     @Min(value = 0, message = "total_money must be >= 0")
     @JsonProperty("total_money")
-    private int totalMoney;
+    private Float totalMoney;
 
     private String color;
 }
